+++ internal/models/maps.go
@@ -1,0 +1,24 @@
+package models
+
+var DriverURLMap = map[string]string{
+	"Max Verstappen":        "max-verstappen",
+	"Lando Norris":          "lando-norris",
+	"Lewis Hamilton":        "lewis-hamilton",
+	"Charles Leclerc":       "charles-leclerc",
+	"George Russell":        "george-russell",
+	"Oscar Piastri":         "oscar-piastri",
+	"Fernando Alonso":       "fernando-alonso",
+	"Lance Stroll":          "lance-stroll",
+	"Pierre Gasly":          "pierre-gasly",
+	"Jack Doohan":           "jack-doohan",
+	"Esteban Ocon":          "esteban-ocon",
+	"Oliver Bearman":        "oliver-bearman",
+	"Alexander Albon":       "alexander-albon",
+	"Carlos Sainz":          "carlos-sainz",
+	"Isack Hadjar":          "isack-hadjar",
+	"Liam Lawson":           "liam-lawson",
+	"Nico Hulkenberg":       "nico-hulkenberg",
+	"Gabriel Bortoleto":     "gabriel-bortoleto",
+	"Andrea Kimi Antonelli": "kimi-antonelli",
+	"Yuki Tsunoda":          "yuki-tsunoda",
+}