+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [0.2.1] - YYYY-MM-DD  
+
+### Added
+- Driver standings now link to bio pages.
+
+### Fixed
+- Tray menu options have been fixed.
+
+---
 
 ## [0.2.0] - 2025-03-30